refactor: `OrderForm` page is now `Order`, extract form into its own component (`OrderForm`)
