+++ application/src/components/index.js
@@ -1,10 +1,11 @@
 import Login from './login/login';
 import Main from './main/main';
 import Nav from './nav/nav';
+import Order from './order/order';
 import OldOrderForm from './old-order-form/orderForm';
 import OrderForm from './order-form/order-form';
 import Template from './common/template';
 import OldViewOrders from './old-view-orders/viewOrders';
 import ViewOrders from './view-orders/viewOrders';
 
-export { Login, OldOrderForm, OrderForm, Main, Nav, Template, OldViewOrders, ViewOrders };
+export { Login, Order, OldOrderForm, OrderForm, Main, Nav, Template, OldViewOrders, ViewOrders };
