+++ application/src/components/order-form/orderForm.css
@@ -3,11 +3,6 @@
     font-size: 2em;
 }
 
-.form-wrapper {
-    margin-left: 5%;
-    margin-top: 2%;
-}
-
 .menu-select {
     margin-bottom: 2%;
     width: 50%;
