+++ application/src/components/order/order.css
@@ -1,0 +1,24 @@
+.form-label {
+  font-family: "Optima";
+  font-size: 2em;
+}
+
+.form-wrapper {
+  margin-left: 5%;
+  margin-top: 2%;
+}
+
+.menu-select {
+  margin-bottom: 2%;
+  width: 50%;
+}
+
+.order-btn {
+  margin-left: 32%;
+}
+
+.qty-label {
+  font-family: "Optima";
+  font-size: 1.5em;
+  margin-right: 1%;
+}
